Updating item_analytics test. WIP
